requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
noisereduce==3.0.0

# Speech-to-text
openai-whisper==20231117
transformers==4.36.0
torch==2.1.1
torchaudio==2.1.1

# NLP and sentiment analysis
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 